refactor/#30: nestJS에 걸맞는 파일이름으로 대변경
