+++ src/users/dtos/get-users-by-animal.dto.ts
@@ -2,7 +2,7 @@ import { IsEnum, IsNumber, IsOptional, IsString } from "class-validator";
 import { Type } from "class-transformer";
 import { Animal } from "../../common/enum/animal.enum";
 import { ApiProperty, ApiPropertyOptional } from "@nestjs/swagger";
-import { orderByField } from "../enum/orderByFeild-enum";
+import { orderByField } from "../enum/orderByFeild.enum";
 
 export class GetUsersByAnimalDto {
   @ApiProperty({
