+++ src/users/users.repository.ts
@@ -10,7 +10,7 @@ import {
 } from "@prisma/client";
 import { JsonValue } from "@prisma/client/runtime/library";
 import { PrismaService } from "src/prisma/prisma.service";
-import { DomainEnum } from "./enum/domain-enum";
+import { DomainEnum } from "./enum/domain.enum";
 
 @Injectable()
 export class UsersRepository {
